Refactor/project

See merge request s07-final/S07P31A606!22

+++ back/src/main/java/cseon/api/repository/WorkbookRepository.java
@@ -14,7 +14,7 @@ public interface WorkbookRepository extends JpaRepository<Workbook, Long> {
     @Query(value = "select w from Workbook w")
     Optional<List<Workbook>> findAllWorkbooks();
 
-    Optional<Workbook> findWorkbookByWorkbookId(int workbookId);
+    Optional<Workbook> findWorkbookByWorkbookId(Long workbookId);
 
     Optional<Workbook> findWorkbooksByWorkbookCreatedByAndWorkbookName(Long workbookCreatedBy, String workbookName);
